WalletApp/requirement.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0